refactor: DELETE original spaCy tokenizer, rename stream_spacy to spacy
